Perf work query and decompression (#7558)
* code review

* fix conflicts

* code review

* code review

* code review

* some cleanup

* restore heuristic packed

* more cleanup

* make init array lighter and test query engine arm

* test flex compression

* confirm heuristic for flex + prune okish benchmarks

* simplify getters for compressed int interners

* specialised find for compressed arrays of int mixed types

* do not compress composite array for mixed

* added the possibility to fetch multiple values in a range b,e for compressed array

* restore all benchmarks

* fetch_all values with unaligned iterator for get_all

* better heuristics in packed

* fix value fetching

* faster decompression

* better heuristic for flex

* fix build windows x86

* fix asan correct report of accessing too much memory passed the last word

* use unaligned_iterator::get for getAll

* fix get_all

* code review

* lint

* fix upper bound for compressed arrays

* reverse logic branch hint

* revert branch hint

+++ src/realm/integer_packed_compressor.cpp
@@ -32,10 +32,10 @@
 
 using namespace realm;
 
-void PackedCompressor::init_array(char* h, uint8_t flags, size_t v_width, size_t v_size) const
+void PackedCompressor::init_array(char* h, uint8_t flags, uint8_t v_width, size_t v_size) const
 {
     using Encoding = NodeHeader::Encoding;
-    init_header((char*)h, Encoding::Packed, flags, v_width, v_size);
+    init_header((char*)h, Encoding::Packed, flags, static_cast<uint8_t>(v_width), v_size);
 }
 
 void PackedCompressor::copy_data(const Array& origin, Array& arr) const
